requirements.txt
# --- Core Backend ---
fastapi
uvicorn[standard]

# --- Supabase + Auth ---
supabase
python-dotenv

# --- AI + Resume Logic ---
openai==1.79.0
PyMuPDF                # Required for fitz PDF parsing
python-docx
fpdf

# --- Data Handling ---
pandas==1.5.3
numpy==1.23.5
